calls_collector now follows requirements of function_patch callbacks
- calls_collector tests refactored not to use penter/pexit hooks;
- the definition of calls_collector global instance is no longer inside calls_collector itself;
- penter/pexit static hooks refactored to use general signature of on_enter/on_exit;
- ExitProcess patcing refactored to use assembler_intel commands.

+++ common/memory_protection.h
@@ -14,7 +14,7 @@ namespace micro_profiler
 
 	private:
 		void *_address;
-		unsigned _size;
+		size_t _size;
 		unsigned _previous_access;
 	};
 }
